requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
celery==5.3.4
redis==5.0.1
django-tenant-schemas==1.11.0
django-guardian==2.4.0
Pillow==10.1.0
boto3==1.29.7
requests==2.31.0
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.5
pandas==2.1.3
numpy==1.26.2
